\name{get_bdd_library}
\alias{get_bdd_library}
\docType{package}
\title{
GETS THE BDD MANAGER LIBRARY
}
\description{
This instruction returns the name of the BDD manager chosen.
}
\usage{
get_bdd_library(string bdd_name)
}
\arguments{
\item{bdd_name}{(Optional) Name of the BDD.}
}
\value{
N/A.
}
\examples{
get_bdd_library()
get_bdd_library("bdd_1")
}
