\name{print_bdd}
\alias{print_bdd}
\docType{package}
\title{
PRINT A SOLVED BDD
}
\description{
This instruction prints the solution of a BDD.
}
\usage{
print_bdd(string bdd_name)
}
\arguments{
\item{bdd_name}{(Optional) Name of the BDD.}
}
\value{
N/A.
}
\examples{
print_bdd("bdd_1")
print_bdd()
}
