\name{print_variables}
\alias{print_variables}
\docType{package}
\title{
PRINTS THE VARIABLES CREATED
}
\description{
This function prints a table showing the index and the content of the variables created.
}
\usage{
print_variables(string bdd_name)
}
\arguments{
\item{bdd_name}{(Optional) Name of the BDD.}
}
\value{
N/A
}
\examples{
print_variables()
print_variables("bdd_1")
}
