\name{rbdd-package}
\alias{rbdd-package}
\alias{rbdd}
\docType{package}
\title{
\packageTitle{rbdd}
}
\description{
\packageDescription{rbdd}
}
\details{

The package allows the use of BDD functions implementing the buddy and cudd libraries.

It has been built using the Rcpp package.

The functions implemented are based on the John Whaley's JavaBDD package
}
\author{
\packageAuthor{rbdd}

Maintainer: \packageMaintainer{rbdd}
}
\references{
\code{\link{http://javabdd.sourceforge.net/}}
\code{\link{http://buddy.sourceforge.net/manual/main.html}}
\code{\link{http://vlsi.colorado.edu/~fabio/CUDD/html/}}
}
\keyword{ package }
\seealso{
\code{help(Rcpp)}
}
